+++ metadata.rb
@@ -8,6 +8,8 @@ license          'Apache-2.0'
 description      'Installs and configures apache2'
 version          '8.9.1'
 
+depends 'yum-epel'
+
 supports 'debian'
 supports 'ubuntu'
 supports 'redhat'